+++ csvkit/cli.py
@@ -167,6 +167,9 @@ class CSVKitUtility(object):
         if 'S' not in self.override_flags:
             self.argparser.add_argument('-S', '--skipinitialspace', dest='skipinitialspace', action='store_true',
                                         help='Ignore whitespace immediately following the delimiter.')
+        if 'blanks' not in self.override_flags:
+            self.argparser.add_argument('--blanks', dest='blanks', action='store_true',
+                                        help='Do not convert "", "na", "n/a", "none", "null", "." to NULL.')
         if 'date-format' not in self.override_flags:
             self.argparser.add_argument('--date-format', dest='date_format',
                                         help='Specify a strptime date format string like "%%m/%%d/%%Y".')
@@ -194,7 +197,7 @@ class CSVKitUtility(object):
                                         help='When interpreting or displaying column numbers, use zero-based numbering instead of the default 1-based numbering.')
 
         self.argparser.add_argument('-V', '--version', action='version', version='%(prog)s 1.0.2',
-                                        help='Display version information and exit.')
+                                    help='Display version information and exit.')
 
     def _open_input_file(self, path):
         """
@@ -284,21 +287,22 @@ class CSVKitUtility(object):
 
     def get_column_types(self):
         if getattr(self.args, 'blanks', None):
-            text_type = agate.Text(cast_nulls=False)
+            type_kwargs = {'null_values': ()}
         else:
-            text_type = agate.Text()
+            type_kwargs = {}
 
-        if self.args.no_inference:
-            return agate.TypeTester(types=[text_type])
-        else:
-            return agate.TypeTester(types=[
-                agate.Boolean(),
-                agate.Number(locale=self.args.locale),
-                agate.TimeDelta(),
-                agate.Date(date_format=self.args.date_format),
-                agate.DateTime(datetime_format=self.args.datetime_format),
-                text_type
-            ])
+        types = [agate.Text(**type_kwargs)]
+
+        if not self.args.no_inference:
+            types = [
+                agate.Boolean(**type_kwargs),
+                agate.Number(locale=self.args.locale, **type_kwargs),
+                agate.TimeDelta(**type_kwargs),
+                agate.Date(date_format=self.args.date_format, **type_kwargs),
+                agate.DateTime(datetime_format=self.args.datetime_format, **type_kwargs),
+            ] + types
+
+        return agate.TypeTester(types=types)
 
     def get_column_offset(self):
         if self.args.zero_based: